Move proofs to separate modules
